pyproject.toml
[project]
name = "scarlett-prompt"
version = "4.0.0"
description = "Kujou Scarlett — Personal Maid-Bot"
requires-python = ">=3.11"
dependencies = [
    "cerebras-cloud-sdk",
    "python-dotenv",
    "rich",
    "mem0ai",
    "sentence-transformers",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sse-starlette>=2.0",
    "httpx>=0.27",
    "msgpack>=1.0",
    "beautifulsoup4>=4.12",
]

[project.scripts]
build-prompt = "build:main"
